api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API for RoaderBot"
authors = ["komediantto <komediantto@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
python-dotenv = "^1.0.0"
aiogram = "^2.25.1"
loguru = "^0.7.0"
requests = "^2.28.2"
python-telegram-bot-pagination = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
